refactor(feedDetail): 뷰 홀더 패키징
